refactor: move all services to client folder
